backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.0
langchain-community==0.3.0
langchain-chroma==0.1.4
langchain-openai==0.2.0
chromadb==0.5.15
psycopg2-binary==2.9.10
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
unstructured[pdf,docx,pptx]==0.16.0
pdfplumber==0.11.0
pypdf==5.0.0
tiktoken==0.8.0
